% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttondown.R
\name{get_api_key}
\alias{get_api_key}
\title{gets a buttondown API key as plaintext from the keychain}
\usage{
get_api_key()
}
\value{
api_key
}
\description{
Note that you may have to give rstudio permission to access your keychain
the first time you run it.
}
