% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttondown.R
\name{get_subscribers}
\alias{get_subscribers}
\title{Get all subscribers as a data.frame}
\usage{
get_subscribers()
}
\value{
data.frame of all subscribers
}
\description{
Get all subscribers as a data.frame
}
\examples{

if(interactive()){
  #make sure your api key is set using set_api_key()
  df <- get_subscribers()
  head(df)
}
}
