% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttondown.R
\name{list_emails}
\alias{list_emails}
\title{List all emails and their status}
\usage{
list_emails()
}
\value{
data.frame of emails with the following columns:
}
\description{
List all emails and their status
}
\examples{

if(interactive()){
  # make sure you've set your api key with set_api_key()
  df <- list_emails()
  head(df)
}
}
