% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttondown.R
\name{send_email}
\alias{send_email}
\title{Sends a draft email from a Rmarkdown file}
\usage{
send_email(input_rmd, draft = FALSE)
}
\arguments{
\item{input_rmd}{- path to an rmarkdown document}

\item{draft}{- status of email. By default, set to `draft`}
}
\value{
response
}
\description{
This function will send a rendered blastula HTML from a markdown file
as a httr2::request() to the buttondown API.
}
\details{
This function differs in that it will send the entire HTML rendered by
blastula, including encoded plots and graphics.

The disadvantage of using this function is that it renders larger HTML files,
and most email clients will cut these off. If that's the case, then use
send_email_with_images() to upload images to the buttondown API and send
a smaller markdown file instead.
}
\examples{
if(interactive()){


}
}
