% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttondown.R
\name{send_email_with_images}
\alias{send_email_with_images}
\title{Render email with images
Given}
\usage{
send_email_with_images(rmarkdown_path)
}
\arguments{
\item{rmarkdown_path}{}
}
\value{
list with the following values:
}
\description{
Render email with images
Given
}
