% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttondown.R
\name{send_image}
\alias{send_image}
\title{Sends an image to the buttondown API}
\usage{
send_image(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
resp - response from buttondown server
}
\description{
Sends an image to the buttondown API
}
