% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttondown.R
\name{send_email}
\alias{send_email}
\title{Sends a draft email from a Rmarkdown file}
\usage{
send_email(input_qmd, draft = FALSE)
}
\arguments{
\item{draft}{}
}
\value{
response
}
\description{
This function will send a rendered blastula HTML from a markdown file
as a httr2::request() to the buttondown API.
}
