% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttondown.R
\name{send_image}
\alias{send_image}
\title{Sends an image to the buttondown API}
\usage{
send_image(filepath)
}
\arguments{
\item{filepath}{- path to an image file}
}
\value{
resp - httr2 response from buttondown server
}
\description{
Sends an image to the buttondown API
}
\examples{
img_file <- system.files("tests/testthat/test_pic.png", package="buttondownr")
resp <- send_image(img_file)
resp
}
